Revert changes from last commit
